refactor stories into separate modules
